services/pipeline_executor.py
import yaml
import docker
from models.pipeline import PipelineExecution, PipelineConfiguration
from app import db

class PipelineExecutor:
    def __init__(self):
        self.docker_client = docker.from_env()
    
    def parse_pipeline_yaml(self, yaml_content):
        """Parse and validate pipeline YAML"""
        try:
            config = yaml.safe_load(yaml_content)
            return config
        except yaml.YAMLError as e:
            raise ValueError(f"Invalid YAML: {str(e)}")
    
    def execute_pipeline(self, execution_id):
        """Execute a pipeline (placeholder implementation)"""
        execution = PipelineExecution.query.get(execution_id)
        if not execution:
            return False
        
        try:
            execution.status = 'running'
            db.session.commit()
            
            # Get YAML content
            config = PipelineConfiguration.query.get(execution.configuration_id)
            yaml_content = config.yaml_content
            
            if config.yaml_source == 'repo':
                # TODO: Fetch YAML from repository
                yaml_content = self.fetch_yaml_from_repo(
                    config.repo_url, 
                    config.repo_branch, 
                    config.yaml_file_path
                )
            
            # Parse pipeline configuration
            pipeline_config = self.parse_pipeline_yaml(yaml_content)
            
            # Execute tasks (simplified)
            logs = []
            for task in pipeline_config.get('tasks', []):
                logs.append(f"Executing: {task['name']}")
                logs.append(f"Command: {task['command']}")
                # TODO: Actually execute the command in Docker container
            
            execution.status = 'success'
            execution.logs = '\n'.join(logs)
            
        except Exception as e:
            execution.status = 'failed'
            execution.error_message = str(e)
        
        finally:
            from datetime import datetime
            execution.completed_at = datetime.utcnow()
            db.session.commit()
        
        return execution.status == 'success'
    
    def fetch_yaml_from_repo(self, repo_url, branch, file_path):
        """Fetch YAML file from Git repository (placeholder)"""
        # TODO: Implement Git repository fetching
        return "env_image: ubuntu:24.10\ntasks:\n  - name: placeholder\n    command: echo 'Hello World'"

