requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
PyYAML==6.0.1
docker==6.1.3
python-dotenv==1.0.0
